requirements/lock.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/lock.txt requirements/dev.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
authlib==1.3.2
    # via -r requirements/base.in
bcrypt==5.0.0
    # via passlib
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
click==8.3.1
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   pytest
cryptography==46.0.3
    # via
    #   authlib
    #   python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.110.0
    # via -r requirements/base.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.27.2
    # via -r requirements/base.in
idna==3.11
    # via
    #   anyio
    #   httpx
imageio==2.37.2
    # via scikit-image
iniconfig==2.3.0
    # via pytest
itsdangerous==2.1.2
    # via -r requirements/base.in
jinja2==3.1.6
    # via -r requirements/base.in
joblib==1.5.3
    # via scikit-learn
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.3
    # via jinja2
networkx==3.6.1
    # via scikit-image
numpy==1.26.4
    # via
    #   -r requirements/base.in
    #   imageio
    #   opencv-python
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
opencv-python==4.11.0.86
    # via -r requirements/base.in
packaging==25.0
    # via
    #   lazy-loader
    #   pytest
    #   scikit-image
passlib[bcrypt]==1.7.4
    # via -r requirements/base.in
pillow==10.4.0
    # via
    #   -r requirements/base.in
    #   imageio
    #   scikit-image
pluggy==1.6.0
    # via pytest
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via fastapi
pydantic-core==2.41.5
    # via pydantic
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via -r requirements/test.in
python-dotenv==1.0.1
    # via -r requirements/base.in
python-jose[cryptography]==3.5.0
    # via -r requirements/base.in
python-multipart==0.0.9
    # via -r requirements/base.in
rsa==4.9.1
    # via python-jose
scikit-image==0.22.0
    # via -r requirements/base.in
scikit-learn==1.4.2
    # via -r requirements/base.in
scipy==1.12.0
    # via
    #   -r requirements/base.in
    #   scikit-image
    #   scikit-learn
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via httpx
starlette==0.36.3
    # via
    #   -r requirements/base.in
    #   fastapi
threadpoolctl==3.6.0
    # via scikit-learn
tifffile==2025.12.20
    # via scikit-image
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
uvicorn==0.27.1
    # via -r requirements/base.in
